+++ TS-web-app/Model/TableauSolution.cs
@@ -5,7 +5,7 @@ namespace TsWebApp.Model {
 
     public class TableauSolution {
 
-        public ulong SolutionId { get; set; }
+        public int SolutionId { get; set; }
         public TableauInput TableauInput { get; set; }
         public SolutionNode SolutionNode { get; set; }
         public DateTime RequestDateTime { get; set; }
